refactor: move unwanted column drop logic to a dedicated file
