+++ etl/transform/transform.py
@@ -1,4 +1,5 @@
 import pandas as pd
+from etl.transform.helper.drop import drop_unwanted_columns
 from etl.transform.helper.clean import clean_data
 from etl.transform.helper.convert import convert_data
 from etl.transform.helper.nulls import handle_null_values
@@ -13,13 +14,3 @@ def transform_data(df) -> pd.DataFrame:
     df_transformed = handle_null_values(df_transformed)
 
     return df_transformed
-
-
-def drop_unwanted_columns(df) -> pd.DataFrame:
-    columns_to_drop = [
-        'ID', 'SSN', 'Num_Bank_Accounts', 'Num_Credit_Card', 'Interest_Rate',
-        'Num_of_Loan', 'Type_of_Loan', 'Delay_from_due_date',
-        'Num_of_Delayed_Payment', 'Changed_Credit_Limit', 'Num_Credit_Inquiries'
-    ]
-    df_dropped = df.drop(columns=columns_to_drop)
-    return df_dropped
